pyproject.toml
[tool.poetry]
name = "association-rules"
version = "0.1.0"
description = "연관 규칙 기반 상품 추천 시스템"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.6"
pandas = "^1.3.0"
python-dateutil = "^2.8.2"
mlxtend = "^0.19.0"
gradio = "^3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
